doc: clarify browser identifiers
The following methods `browser_process.options` and
`browser_process.spawn` and `browser_process.type` accept either an
absolute browser executable path or the name of a browser. This is not
reflected to clear in the documentation.

This re-phrases it to refer to "browser identifiers", which means either
a browser executable path or the short name of a browser.