lib: make detect yield executable paths only
Detect yields some browser information, but sine it won't yield version
information it makes more sense to have it use executable paths instead
of objects. Its simpler to convey and fits better with `type` etc.

This changes detect to invoke the callback with an array of commands
intead of an array of objects.