test: parameterize tests via env variables
There's a lot of looping going on in the tests, which makes them very
tedious to maintain, there's no ability to control which browsers to
expect on a given test platform.

This removes all the iteration over different browsers, instead we use
an environment variable to control which browser is being tested and set
up our test matrix via a makefile. Doing this via make also allows make
to be called with multi-threading (e.g `make -j`).