test: seperate different browser type tests
When testing the actual expected options or spawn behavior, chrome,
firefox and electron have little in common. So we end up having branches
trying to accomodate them all in the same test.

This splits them out into multiple tests and uses `skip` on tape to skip
the test when it is not relevant.