make: use shallow permalinks
Deeply nested permalinks gave a good structure, but there were just too
many edge cases for what it's worth when it comes to resolving relative
urls. Without resolving to plugins we really need to keep the same basic
structure as the source documents are.

This reverts to a basic readme -> index, .md -> .html substitution for
permalinks.