lib: make find deterministic
Currently `find` will return the first match it finds, which means its
basically a race condition and won't always return the same result. This
is not what we want, we need it to always return the first match in the
order the paths are listed.

This makes it deterministic by checking all the paths, and returning
the first match sorted by index.