lib: resolve executable paths with fs.realpath
Currently `spawn` resolves symlinks with fs.realpath, this is a bit
difficult to test since it is inside the function. Better to opt for the
more debuggable option and have the paths be resolved when the search is
done.

This resolves executable paths in `detect`, `detectSync`, `find` and
`findSync` with `fs.realpath`.